Deleted the old index.html file
I deleted it and updated it to the actual file we are working on. Only work on the index.html. If you need to create more files for reasons such as trying out experimental codes or creating a function and soever, please create as many as you want. You can create a new folder in the repository if you want, anything so its intuitive and easy to access.